requirements-telemetry.txt
# OpenTelemetry Core
opentelemetry-api
opentelemetry-sdk

# OTLP Exporter
opentelemetry-exporter-otlp
opentelemetry-exporter-otlp-proto-grpc

# Programmatic Instrumentation
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-sqlalchemy